web: wip module page

+++ Roki.Web/Services/DiscordService.cs
@@ -29,7 +29,7 @@ namespace Roki.Web.Services
             var guilds = new List<DiscordGuild>();
             foreach (JsonElement element in json.Result.RootElement.EnumerateArray())
             {
-                if (element.GetProperty("owner").GetBoolean() || (uint.Parse(element.GetProperty("permissions").GetString()!) & 0x20) == 0x20)
+                if (element.GetProperty("owner").GetBoolean() || (ulong.Parse(element.GetProperty("permissions").GetString()!) & 0x20) == 0x20)
                 {
                     guilds.Add(new DiscordGuild
                     {
